+++ packages/jest-haste-map/src/index.ts
@@ -16,7 +16,7 @@ import {Stats, readFileSync, writeFileSync} from 'graceful-fs';
 import type {Config} from '@jest/types';
 import {escapePathForRegex} from 'jest-regex-util';
 import {requireOrImportModule} from 'jest-util';
-import {Worker} from 'jest-worker';
+import {JestWorkerFarm, Worker} from 'jest-worker';
 import HasteFS from './HasteFS';
 import HasteModuleMap from './ModuleMap';
 import H from './constants';
@@ -108,13 +108,16 @@ type Watcher = {
   close(): Promise<void>;
 };
 
-type WorkerInterface = {worker: typeof worker; getSha1: typeof getSha1};
+type HasteWorker = typeof import('./worker');
 
-export {default as ModuleMap} from './ModuleMap';
-export type {SerializableModuleMap} from './types';
-export type {IModuleMap} from './types';
 export type {default as FS} from './HasteFS';
-export type {ChangeEvent, HasteMap as HasteMapObject} from './types';
+export {default as ModuleMap} from './ModuleMap';
+export type {
+  ChangeEvent,
+  HasteMap as HasteMapObject,
+  IModuleMap,
+  SerializableModuleMap,
+} from './types';
 
 const CHANGE_INTERVAL = 30;
 const MAX_WAIT_TIME = 240000;
@@ -216,7 +219,7 @@ export default class HasteMap extends EventEmitter {
   private _isWatchmanInstalledPromise: Promise<boolean> | null = null;
   private _options: InternalOptions;
   private _watchers: Array<Watcher>;
-  private _worker: WorkerInterface | null;
+  private _worker: JestWorkerFarm<HasteWorker> | HasteWorker | null;
 
   static getStatic(config: Config.ProjectConfig): HasteMapStatic {
     if (config.haste.hasteMapModulePath) {
@@ -690,7 +693,7 @@ export default class HasteMap extends EventEmitter {
       this._recoverDuplicates(hasteMap, relativeFilePath, fileMetadata[H.ID]);
     }
 
-    const promises = [];
+    const promises: Array<Promise<void>> = [];
     for (const relativeFilePath of filesToProcess.keys()) {
       if (
         this._options.skipPackageJson &&
@@ -726,9 +729,7 @@ export default class HasteMap extends EventEmitter {
   private _cleanup() {
     const worker = this._worker;
 
-    // @ts-expect-error
-    if (worker && typeof worker.end === 'function') {
-      // @ts-expect-error
+    if (worker && 'end' in worker && typeof worker.end === 'function') {
       worker.end();
     }
 
@@ -745,19 +746,20 @@ export default class HasteMap extends EventEmitter {
   /**
    * Creates workers or parses files and extracts metadata in-process.
    */
-  private _getWorker(options?: {forceInBand: boolean}): WorkerInterface {
+  private _getWorker(options?: {
+    forceInBand: boolean;
+  }): JestWorkerFarm<HasteWorker> | HasteWorker {
     if (!this._worker) {
       if ((options && options.forceInBand) || this._options.maxWorkers <= 1) {
         this._worker = {getSha1, worker};
       } else {
-        // @ts-expect-error: assignment of a worker with custom properties.
         this._worker = new Worker(require.resolve('./worker'), {
           exposedMethods: ['getSha1', 'worker'],
           // @ts-expect-error: option does not exist on the node 12 types
           forkOptions: {serialization: 'json'},
           maxRetries: 3,
           numWorkers: this._options.maxWorkers,
-        }) as WorkerInterface;
+        }) as JestWorkerFarm<HasteWorker>;
       }
     }